updated after FRS app weather data testing
